requirements.txt
streamlit
ta
pandas
numpy
python-binance
plotly
